Cleanup metrics
